+++ save-agent/src/commonMain/kotlin/com/saveourtool/save/agent/Main.kt
@@ -17,11 +17,8 @@ import io.ktor.client.plugins.*
 import io.ktor.client.plugins.contentnegotiation.*
 import io.ktor.client.plugins.logging.*
 import io.ktor.serialization.kotlinx.json.*
-import okio.FileSystem
 import okio.Path.Companion.toPath
-import platform.posix.*
 
-import kotlinx.cinterop.staticCFunction
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
 import kotlinx.serialization.ExperimentalSerializationApi
@@ -60,10 +57,7 @@ fun main() {
     logType.set(if (config.debug) LogType.ALL else LogType.WARN)
     logDebugCustom("Instantiating save-agent version $SAVE_CLOUD_VERSION with config $config")
 
-    signal(SIGTERM, staticCFunction<Int, Unit> {
-        logInfoCustom("Agent is shutting down because SIGTERM has been received")
-        exit(1)
-    })
+    catchSigterm()
 
     val httpClient = configureHttpClient(config)
