+++ save-agent/src/commonMain/kotlin/com/saveourtool/save/agent/utils/Logging.kt
@@ -6,13 +6,7 @@
 
 package com.saveourtool.save.agent.utils
 
-import com.saveourtool.save.core.logging.logDebug
-import com.saveourtool.save.core.logging.logError
-import com.saveourtool.save.core.logging.logInfo
 import io.ktor.client.plugins.logging.*
-
-import platform.linux.__NR_gettid
-import platform.posix.syscall
 
 internal val ktorLogger = object : Logger {
     override fun log(message: String) {
@@ -20,14 +14,8 @@ internal val ktorLogger = object : Logger {
     }
 }
 
-fun logErrorCustom(msg: String) = logError(
-    "[tid ${syscall(__NR_gettid.toLong())}] $msg"
-)
+expect fun logErrorCustom(msg: String)
 
-fun logInfoCustom(msg: String) = logInfo(
-    "[tid ${syscall(__NR_gettid.toLong())}] $msg"
-)
+expect fun logInfoCustom(msg: String)
 
-fun logDebugCustom(msg: String) = logDebug(
-    "[tid ${syscall(__NR_gettid.toLong())}] $msg"
-)
+expect fun logDebugCustom(msg: String)
