+++ save-agent/src/jvmMain/kotlin/com/saveourtool/save/agent/utils/PlatformUtils.kt
@@ -10,3 +10,9 @@ actual class GenericAtomicReference<T> actual constructor(valueToStore: T) {
         holder.set(newValue)
     }
 }
+
+internal actual fun getenv(envName: String): String? = System.getenv(envName)
+
+internal actual fun catchSigterm() {
+
+}
