+++ save-agent/src/linuxX64Main/kotlin/com/saveourtool/save/agent/utils/Logging.kt
@@ -14,20 +14,14 @@ import io.ktor.client.plugins.logging.*
 import platform.linux.__NR_gettid
 import platform.posix.syscall
 
-internal val ktorLogger = object : Logger {
-    override fun log(message: String) {
-        logInfoCustom("[HTTP Client] $message")
-    }
-}
-
-fun logErrorCustom(msg: String) = logError(
+actual fun logErrorCustom(msg: String) = logError(
     "[tid ${syscall(__NR_gettid.toLong())}] $msg"
 )
 
-fun logInfoCustom(msg: String) = logInfo(
+actual fun logInfoCustom(msg: String) = logInfo(
     "[tid ${syscall(__NR_gettid.toLong())}] $msg"
 )
 
-fun logDebugCustom(msg: String) = logDebug(
+actual fun logDebugCustom(msg: String) = logDebug(
     "[tid ${syscall(__NR_gettid.toLong())}] $msg"
 )
